+++ tests/test_cli.py
@@ -16,6 +16,8 @@ from typing import Iterator
 
 from unittest import mock
 
+from libcst import parse_module
+from libcst.codemod.visitors import ImportItem
 
 from monkeytype import cli
 from monkeytype.config import DefaultConfig
@@ -450,3 +452,69 @@ def test_apply_stub_using_libcst__overwrite_existing_annotations():
         textwrap.dedent(source),
         overwrite_existing_annotations=True,
     ) == textwrap.dedent(expected)
+
+
+def test_apply_stub_using_libcst__confine_new_imports_in_type_checking_block():
+    source = """
+        def spoof(x):
+            return x.get_some_object()
+    """
+    stub = """
+        from some.module import (
+            AnotherObject,
+            SomeObject,
+        )
+
+        def spoof(x: AnotherObject) -> SomeObject: ...
+    """
+    expected = """
+        from __future__ import annotations
+        from typing import TYPE_CHECKING
+
+        if TYPE_CHECKING:
+            from some.module import AnotherObject, SomeObject
+
+        def spoof(x: AnotherObject) -> SomeObject:
+            return x.get_some_object()
+    """
+
+    assert cli.apply_stub_using_libcst(
+        textwrap.dedent(stub),
+        textwrap.dedent(source),
+        overwrite_existing_annotations=True,
+        confine_new_imports_in_type_checking_block=True,
+    ) == textwrap.dedent(expected)
+
+
+def test_get_newly_imported_items():
+    source = """
+        import q
+        from x import Y
+    """
+    stub = """
+        from a import (
+            B,
+            C,
+        )
+        import d
+        import q, w, e
+        from x import (
+            Y,
+            Z,
+        )
+        import z as t
+    """
+    expected = {
+        ImportItem('a', 'B'),
+        ImportItem('a', 'C'),
+        ImportItem('d'),
+        ImportItem('w'),
+        ImportItem('e'),
+        ImportItem('x', 'Z'),
+        ImportItem('z', None, 't'),
+    }
+
+    assert expected == set(cli.get_newly_imported_items(
+        parse_module(textwrap.dedent(stub)),
+        parse_module(textwrap.dedent(source)),
+    ))